tcc_scraper/requirements.txt
evohomeclient
influxdb-client
